+++ src/TimePanel.h
@@ -6,8 +6,8 @@ class TimePanel : public Panel
 {
 private:	
 	std::string _text;
-	int _timeAvailable;
-	int _t0;
+	size_t _timeAvailable;
+	size_t _t0;
 	int _textX;
 	int _textY;
 public:
